Prevent application from retrying endpoint
It appears that when a Yelp user image is missing, although we include
an 'onerror' HTML attribute to display a default image, it still seems to
trigger a second hit to the translation endpoint. Since this is not
initiated by the user from selecting a language from the dropdown menu,
a target language is not passed in.

If "None" is passed in instead of a target language, then the user is
not attempting to hit the endpoint, and we simply return to avoid
producing an error.